organize v2, v3 instrumentation
