+++ index.js
@@ -11,24 +11,24 @@
  * then the supportability metrics for custom instrumentation will trigger.
  */
 const newrelic = require('newrelic')
-newrelic.instrumentConglomerate('aws-sdk', require('./lib/instrumentation'))
+newrelic.instrumentConglomerate('aws-sdk', require('./lib/v2/instrumentation'))
 newrelic.instrument({
   moduleName: '@aws-sdk/smithy-client',
-  onResolved: require('./lib/smithy-client')
+  onResolved: require('./lib/v3/smithy-client')
 })
 newrelic.instrumentMessages({
   moduleName: '@aws-sdk/client-sns',
-  onResolved: require('./lib/v3-sns')
+  onResolved: require('./lib/v3/sns')
 })
 newrelic.instrumentMessages({
   moduleName: '@aws-sdk/client-sqs',
-  onResolved: require('./lib/v3-sqs')
+  onResolved: require('./lib/v3/sqs')
 })
 newrelic.instrumentDatastore({
   moduleName: '@aws-sdk/client-dynamodb',
-  onResolved: require('./lib/v3-client-dynamodb')
+  onResolved: require('./lib/v3/client-dynamodb')
 })
 newrelic.instrumentDatastore({
   moduleName: '@aws-sdk/lib-dynamodb',
-  onResolved: require('./lib/v3-dynamodb-doc-client')
+  onResolved: require('./lib/v3/dynamodb-doc-client')
 })
