+++ lib/v2/core.js
