+++ lib/v2/instrumentation-helper.js
