+++ lib/v2/sns.js
