+++ lib/v3/client-dynamodb.js
