+++ lib/v3/dynamodb-doc-client.js
