+++ tests/versioned/v2/aws-sdk.tap.js
@@ -32,7 +32,7 @@ tap.test('aws-sdk', (t) => {
     helper.registerInstrumentation({
       moduleName: 'aws-sdk',
       type: 'conglomerate',
-      onRequire: require('../../../lib/instrumentation')
+      onRequire: require('../../../lib/v2/instrumentation')
     })
     AWS = require('aws-sdk')
     AWS.config.update({ region: 'us-east-1' })
