+++ tests/versioned/v2/dynamodb.tap.js
@@ -34,7 +34,7 @@ tap.test('DynamoDB', (t) => {
     helper.registerInstrumentation({
       moduleName: 'aws-sdk',
       type: 'conglomerate',
-      onRequire: require('../../../lib/instrumentation')
+      onRequire: require('../../../lib/v2/instrumentation')
     })
 
     AWS = require('aws-sdk')
