+++ tests/versioned/v2/http-services.tap.js
@@ -31,7 +31,7 @@ tap.test('AWS HTTP Services', (t) => {
     helper.registerInstrumentation({
       moduleName: 'aws-sdk',
       type: 'conglomerate',
-      onRequire: require('../../../lib/instrumentation')
+      onRequire: require('../../../lib/v2/instrumentation')
     })
     AWS = require('aws-sdk')
     AWS.config.update({ region: 'us-east-1' })
