+++ tests/versioned/v3/lambda.tap.js
