+++ tests/versioned/v3/redshift.tap.js
