+++ tests/versioned/v3/rekognition.tap.js
